initial commit
The RocketBook app supports a good selection of cloud destination, but for policy reasons, we can't all necessarily use them. So this script, allows you to use the basic email method but automatically sort them in a folder structure on your mac.

This script is run by a rule in Mail.app, so need to be copied to ~/Library/Application Scripts/com.apple.mail/ and then specified in a mail rule of action type run AppleScript. The name of the rule specifies the folder (relative to the logged in user's home folder) in which to save the attachments. It should be defined in AppleScript notation i.e. "My Folder:RocketBook". 

Within the 'RocketBook' folder it will create a sub folder for each notebook section as denoted by a 'tag' in the recipient address.
e.g. if you mail address is myname@maildomain.org then you might set the RocketBook app to the 'apple' section to myname+apple@maildomain.org and the script will create a folder 'apple' and save the relevant attachments in there. 

Of course this is more useful if the 'apple' image denotes your lecture notes for 'physics', then you might set the address 'myname+physics@maildomain.org', that way when you re-assign the 'apple' to represent 'gardening' then you update the target address and your physics and gardening note will be kept separate.

It will run only on the mails that match the rule, where the message has a status of unread. After saving the attachment(s) it will set the status of the message to read (so it doesn't get picked up on the next run.